requirements-dev.txt
# Optional development-only Python dependencies
# Install with: pip install -r requirements.txt -r requirements-dev.txt

django-debug-toolbar>=4.0.0
django-extensions>=3.2.0
pylint-django>=2.5.0
